[PATCH] Fix USB address setting
You wouldn't expect that an innocent two-line patch would practically
destroy the entire functionality of the USB subsystem.  But that's what my
last patch did.  I didn't realize until I saw it in Greg's list of patches
forwarded to Linus; that patch includes a mistakenly unreverted prior
change.  This explains the problems I was having yesterday, and probably
also the problems a great many people are having today.

Anyway, here's the correction -- it puts things back to what they should
have been in the first place.  Greg, please apply this and send it to
Linus for his tree as soon as reasonably possible.