[PATCH] Fix mem= options
Patch from "Randy.Dunlap" <rddunlap@osdl.org>

Reverts the recent alteration of the format of the `mem=' option.  This is
because `mem=' is interpreted by bootloaders and may not be freely changed.

Instead, the new functionality to set specific memory region usages is
provided via the new "memmap=" option.

The documentation for memmap= is added, and the documentation for mem= is
updated.