[PCMCIA] pcmcia-10: Make cardbus use the new PCI functionality.
Now that we have the critical PCI changes in place, we can convert
cardbus to use this PCI functionality.  This allows us to scan
behind PCI to PCI bridges on cardbus cards, and setup the bus
resources using the generic PCI support code.

Note that drivers/pci/setup-bus.c needs to be built when hotplug
(ie, cardbus) is enabled.