[PATCH] fix fs->lock deadlock with emulated name lookup
The patch below is needed to avoid a deadlock on fs->lock.  Without
the patch, if __emul_lookup_dentry() returns 0, we fail to reacquire
current->fs->lock and then go ahead to read_unlock() it anyhow.  Bad
for your health.

I believe the bug was introduced when the fast pathwalk was reverted
in order to introduce the RCU lockless path walking.